refactor: rename trail interface file
