+++ src/types/Trail.ts
